+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Used named entity-encoding on the server-side
 * `Function.parallel` and `Function.series` will now reject if there is an error in the callback, instead of silently failing
 * Added `Object.isIterable(obj)` to check for an iterable object
+* `RURL.isUrl()` should no longer return true for objects containing just a `href` property
 
 ## 0.7.4 (2020-10-22)